+++ test/utils/Memory.t.sol
@@ -15,10 +15,10 @@ contract MemoryTest is Test {
     // - moving the free memory pointer to far causes OOG errors
     uint256 constant END_PTR = type(uint24).max;
 
-    function testGetsetFreeMemoryPointer(uint256 seed) public pure {
+    function testGetSetFreeMemoryPointer(uint256 seed) public pure {
         bytes32 ptr = bytes32(bound(seed, START_PTR, END_PTR));
-        ptr.asPointer().setFreeMemoryPointer();
-        assertEq(Memory.getFreeMemoryPointer().asBytes32(), ptr);
+        Memory.Pointer.wrap(ptr).unsafeSetFreeMemoryPointer();
+        assertEq(Memory.Pointer.unwrap(Memory.getFreeMemoryPointer()), ptr);
     }
 
     function testAsSliceToBytes(bytes memory input) public pure {